Shape classes are transfered to io's core from ds9's core
